refactor: 탭 관련 api는 ROLE_ADMIN이 아니면 예외를 던지도록 수정

+++ module-community/src/main/java/com/community/board/controller/CommunityTabController.java
@@ -3,12 +3,15 @@ package com.community.board.controller;
 import com.community.board.service.CommunityTabService;
 import com.community.board.dto.CommunityTabDto;
 import com.community.board.entity.CommunityTab;
+import dto.AuthedUserDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @RequestMapping("community/tabs")
@@ -25,19 +28,28 @@ public class CommunityTabController {
     }
 
     @PostMapping
-    public CommunityTabDto.Response saveTab(@RequestBody CommunityTabDto.Request dto) {
+    public CommunityTabDto.Response saveTab(@AuthenticationPrincipal AuthedUserDto user, @RequestBody CommunityTabDto.Request dto) {
+        if (!Objects.equals(user.getRole(), "ROLE_ADMIN")) {
+            throw new IllegalArgumentException("요청 권한이 없습니다.");
+        }
         CommunityTab communityTab = communityTabService.saveTab(dto);
         return CommunityTabDto.from(communityTab);
     }
 
     @PatchMapping
-    public CommunityTabDto.Response updateTab(@RequestBody CommunityTabDto.Update dto) {
+    public CommunityTabDto.Response updateTab(@AuthenticationPrincipal AuthedUserDto user, @RequestBody CommunityTabDto.Update dto) {
+        if (!Objects.equals(user.getRole(), "ROLE_ADMIN")) {
+            throw new IllegalArgumentException("요청 권한이 없습니다.");
+        }
         CommunityTab communityTab = communityTabService.updateTab(dto);
         return CommunityTabDto.from(communityTab);
     }
 
     @DeleteMapping
-    public ResponseEntity<HttpStatus> deleteTab(@RequestBody CommunityTabDto.Delete dto) {
+    public ResponseEntity<HttpStatus> deleteTab(@AuthenticationPrincipal AuthedUserDto user, @RequestBody CommunityTabDto.Delete dto) {
+        if (!Objects.equals(user.getRole(), "ROLE_ADMIN")) {
+            throw new IllegalArgumentException("요청 권한이 없습니다.");
+        }
         communityTabService.deleteTab(dto);
         return ResponseEntity.ok().body(HttpStatus.OK);
     }
